otorisasi admin, user dan guest

+++ foodstore-server/app/category/controller.js
@@ -1,4 +1,5 @@
 const Category = require('./model');
+const { policyFor } = require('../policy');
 
 async function index(req, res, next) {
 
@@ -23,6 +24,14 @@ async function index(req, res, next) {
 
 async function store(req, res, next) {
     try {
+
+        let policy = policyFor(req.user);
+        if (!policy.can('create', 'Category')) { // <-- can create Category
+            return res.json({
+                error: 1,
+                message: `Anda tidak memiliki akses untuk membuat kategori`
+            });
+        }
         // (1) tangkap payload dari _client request_ 
         let payload = req.body;
 
@@ -52,6 +61,15 @@ async function store(req, res, next) {
 
 async function update(req, res, next) {
     try {
+        //--- cek policy ---/
+        let policy = policyFor(req.user);
+        if (!policy.can('update', 'Category')) { // <-- can update Category
+            return res.json({
+                error: 1,
+                message: `Anda tidak memiliki akses untuk mengupdate
+kategori`
+            });
+        }
         // (1) tangkap payload dari _client request_ 
         let payload = req.body;
 
@@ -81,6 +99,14 @@ async function update(req, res, next) {
 
 async function destroy(req, res, next) {
     try {
+        //--- cek policy ---/
+        let policy = policyFor(req.user);
+        if (!policy.can('delete', 'Category')) { // <-- can delete Category
+            return res.json({
+                error: 1,
+                message: `Anda tidak memiliki akses untuk menghapus kategori`
+            });
+        }
 
         // (1) cari dan hapus categori di MongoDB berdasarkan field _id
         let deleted = await Category.findOneAndDelete({
